+++ src/components/FlawForm/FlawForm.vue
@@ -21,6 +21,7 @@ import CweSelector from '@/components/CweSelector/CweSelector.vue';
 import FlawLabelsTable from '@/components/FlawLabels/FlawLabelsTable.vue';
 
 import { useFlawModel } from '@/composables/useFlawModel';
+import { useFlawCvssScores } from '@/composables/useFlawCvssScores';
 
 import LabelTextarea from '@/widgets/LabelTextarea/LabelTextarea.vue';
 import LabelStatic from '@/widgets/LabelStatic/LabelStatic.vue';
@@ -62,27 +63,30 @@ const {
   flawAcknowledgments,
   flawIncidentStates,
   flawReferences,
-  flawRhCvss3,
-  highlightedNvdCvss3String,
   internalComments,
   internalCommentsAvailable,
   isLoadingInternalComments,
   isSaving,
   isValid,
   loadInternalComments,
-  nvdCvss3String,
   osimLink,
   privateComments,
   publicComments,
-  rhCvss3String,
   saveAcknowledgments,
   saveReferences,
   shouldCreateJiraTask,
-  shouldDisplayEmailNistForm,
   systemComments,
   toggleShouldCreateJiraTask,
   updateFlaw,
 } = useFlawModel(props.flaw, onSaveSuccess);
+
+const {
+  flawRhCvss,
+  highlightedNvdCvssString,
+  nvdCvssString,
+  rhCvssString,
+  shouldDisplayEmailNistForm,
+} = useFlawCvssScores();
 
 const { draftFlaw } = useDraftFlawStore();
 let initialFlaw: ZodFlawType;
@@ -243,19 +247,17 @@ const createdDate = computed(() => {
               :withBlank="true"
             />
             <CvssCalculator
-              :id="flawRhCvss3.uuid"
-              v-model:cvss-vector="flawRhCvss3.vector"
-              v-model:cvss-score="flawRhCvss3.score"
+              :id="flawRhCvss.uuid"
             />
             <CvssSection
-              :highlightedNvdCvss3String
+              :highlightedNvdCvssString
               :shouldDisplayEmailNistForm
               :cveId="flaw.cve_id"
               :summary="flaw.comment_zero"
               :bugzilla="bugzillaLink"
-              :cvss="rhCvss3String"
+              :cvss="rhCvssString"
               :allCvss="flaw.cvss_scores"
-              :nistCvss="nvdCvss3String"
+              :nistCvss="nvdCvssString"
             />
             <CweSelector
               v-model="flaw.cwe_id"