+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # OSIM Changelog
 
+## [Unreleased]
+### Fixed
+* Fixed inconsistent local storage cached values (`OSIDB-4129`)
+
 ## [2025.3.2]
 ### Added
 * Added colors to flaw list labels (`OSIDB-3992`)